Work on warnings
