+++ jpype/_jcollection.py
@@ -20,7 +20,7 @@ from collections.abc import Mapping
 import _jpype
 from . import _jclass
 from . import _jcustomizer
-from . import _jtypes
+from . import types as _jtypes
 
 JOverride = _jclass.JOverride
