+++ test/jpypetest/test_properties.py
@@ -1,5 +1,4 @@
 import subrun
-import os
 import unittest
 import jpype
